+++ PSFramework/internal/tepp/scripts/config-fullname.ps1
@@ -1,3 +1,3 @@
 ﻿Register-PSFTeppScriptblock -Name "PSFramework-config-fullname" -ScriptBlock {
 	[PSFramework.Configuration.ConfigurationHost]::Configurations.Values | Where-Object { -not $_.Hidden } | Select-Object -ExpandProperty FullName
-}
+} -Global