+++ PSFramework/internal/tepp/scripts/config-schema.ps1
@@ -1,3 +1,3 @@
 ﻿Register-PSFTeppScriptblock -Name 'PSFramework-Config-Schema' -ScriptBlock {
 	[PSFramework.Configuration.ConfigurationHost]::Schemata.Keys
-}
+} -Global