+++ PSFramework/internal/tepp/scripts/config-validation.ps1
@@ -1,1 +1,3 @@
-﻿Register-PSFTeppScriptblock -Name 'PSFramework-config-validation' -ScriptBlock { [PSFramework.Configuration.ConfigurationHost]::Validation.Keys }
+﻿Register-PSFTeppScriptblock -Name 'PSFramework-config-validation' -ScriptBlock {
+	[PSFramework.Configuration.ConfigurationHost]::Validation.Keys
+} -Global