+++ PSFramework/internal/tepp/scripts/logging-provider.ps1
@@ -1,10 +1,10 @@
 ﻿Register-PSFTeppScriptblock -Name 'PSFramework-logging-provider' -ScriptBlock {
 	(Get-PSFLoggingProvider).Name
-}
+} -Global
 
 Register-PSFTeppScriptblock -Name 'PSFramework-logging-instance-provider' -ScriptBlock {
 	(Get-PSFLoggingProviderInstance).Provider.Name | Select-Object -Unique
-}
+} -Global
 
 Register-PSFTeppScriptblock -Name 'PSFramework-logging-instance-name' -ScriptBlock {
 	if ($fakeBoundParameters.ProviderName)
@@ -12,4 +12,4 @@ Register-PSFTeppScriptblock -Name 'PSFramework-logging-instance-name' -ScriptBlo
 		return (Get-PSFLoggingProviderInstance -ProviderName $fakeBoundParameters.ProviderName).Name
 	}
 	(Get-PSFLoggingProviderInstance).Name | Select-Object -Unique
-}
+} -Global