+++ PSFramework/internal/tepp/scripts/runspace.ps1
@@ -1,3 +1,3 @@
 ﻿Register-PSFTeppScriptblock -Name 'PSFramework-runspace-name' -ScriptBlock {
 	(Get-PSFRunspace).Name
-}
+} -Global